move multiplicity to out of mocular-schema
