+++ experimental/dds/tree2/src/feature-libraries/typed-schema/schemaCollection.ts
@@ -5,7 +5,6 @@
 
 import { assert } from "@fluidframework/core-utils";
 import { Adapters, TreeAdapter, TreeNodeSchemaIdentifier } from "../../core";
-import { Multiplicity } from "../modular-schema";
 import { capitalize, fail, requireAssignableTo } from "../../util";
 import { defaultSchemaPolicy, FieldKinds } from "../default-schema";
 import {
@@ -20,6 +19,7 @@ import {
 } from "./typedTreeSchema";
 import { normalizeFlexListEager } from "./flexList";
 import { Sourced } from "./view";
+import { Multiplicity } from "../multiplicity";
 
 // TODO: tests for this file
